requirements.txt
Pillow~=8.1.0
numpy~=1.20.1
matplotlib~=3.3.4
selenium~=3.141.0